+++ saleor/graphql/media/tests/mutations/test_media_delete.py
@@ -11,10 +11,14 @@ from .....graphql.tests.utils import (
 from .....product import MediaOwnerTypes
 from .....product.error_codes import MediaDeleteErrorCode
 from .....product.media import (
-    MEDIA_OWNER_PERMISSION_MAP,
     OWNER_TYPE_TO_MEDIA_GRAPHQL_TYPE,
 )
 from .....product.models import ProductMedia
+from ..utils import (
+    MEDIA_AUTH_CASES,
+    MEDIA_AUTH_PARAMS,
+    media_global_id,
+)
 
 MEDIA_DELETE_MUTATION = """
     mutation deleteMedia($id: ID!) {
@@ -35,12 +39,6 @@ MEDIA_DELETE_MUTATION = """
 ALL_OWNER_TYPES = MediaOwnerTypes.ALL
 
 
-def _media_global_id(owner_type, media):
-    return graphene.Node.to_global_id(
-        OWNER_TYPE_TO_MEDIA_GRAPHQL_TYPE[owner_type], media.pk
-    )
-
-
 @pytest.mark.parametrize("owner_type", ALL_OWNER_TYPES)
 @patch("saleor.plugins.manager.PluginsManager.media_deleted")
 def test_delete(
@@ -56,7 +54,7 @@ def test_delete(
         permission_manage_products, permission_manage_pages
     )
     media = media_owner.media.create(alt="alt")
-    media_id = _media_global_id(owner_type, media)
+    media_id = media_global_id(owner_type, media)
     variables = {"id": media_id}
 
     # when
@@ -90,31 +88,7 @@ def test_delete_rejects_unknown_media(staff_api_client, permission_manage_pages)
 
 
 @pytest.mark.parametrize("owner_type", ALL_OWNER_TYPES)
-@pytest.mark.parametrize(
-    ("_case", "client_fixture", "permission_fixture", "is_allowed"),
-    [
-        ("Unauthenticated user should be rejected", "api_client", None, False),
-        ("Unprivileged user should be rejected", "user_api_client", None, False),
-        (
-            "Staff user without the permission should be rejected",
-            "staff_api_client",
-            None,
-            False,
-        ),
-        (
-            "Staff user with the owner's permission should be allowed",
-            "staff_api_client",
-            "owner",
-            True,
-        ),
-        (
-            "Staff user with the other domain's permission should be rejected",
-            "staff_api_client",
-            "other",
-            False,
-        ),
-    ],
-)
+@pytest.mark.parametrize(MEDIA_AUTH_PARAMS, MEDIA_AUTH_CASES)
 @patch("saleor.plugins.manager.PluginsManager.media_deleted")
 def test_delete_authorization(
     mock_media_deleted,
@@ -125,26 +99,14 @@ def test_delete_authorization(
     owner_type,
     media_owner,
     request,
+    grant_media_permission,
 ):
     # given
     client = request.getfixturevalue(client_fixture)
-    owner_permission = MEDIA_OWNER_PERMISSION_MAP[owner_type]
-    other_permission = next(
-        permission
-        for permission in MEDIA_OWNER_PERMISSION_MAP.values()
-        if permission != owner_permission
-    )
-    if permission_fixture:
-        permission = {"owner": owner_permission, "other": other_permission}[
-            permission_fixture
-        ]
-        codename = permission.value.split(".")[1]
-        perm_object = request.getfixturevalue(f"permission_{codename}")
-        if client.user:
-            client.user.user_permissions.add(perm_object)
+    grant_media_permission(client, permission_fixture)
 
     media = media_owner.media.create(alt="alt")
-    variables = {"id": _media_global_id(owner_type, media)}
+    variables = {"id": media_global_id(owner_type, media)}
 
     # when
     response = client.post_graphql(MEDIA_DELETE_MUTATION, variables)
@@ -161,3 +123,36 @@ def test_delete_authorization(
         assert content["data"]["mediaDelete"] is None
         assert ProductMedia.objects.filter(pk=media.pk).exists() is True
         mock_media_deleted.assert_not_called()
+
+
+@pytest.mark.parametrize("owner_type", ALL_OWNER_TYPES)
+@patch("saleor.plugins.manager.PluginsManager.product_media_deleted")
+@patch("saleor.plugins.manager.PluginsManager.media_deleted")
+def test_delete_also_fires_the_deprecated_product_media_event(
+    mock_media_deleted,
+    mock_product_media_deleted,
+    owner_type,
+    media_owner,
+    staff_api_client,
+    permission_manage_products,
+    permission_manage_pages,
+):
+    """`PRODUCT_MEDIA_DELETED` must keep firing for product-owned media."""
+    # given
+    staff_api_client.user.user_permissions.add(
+        permission_manage_products, permission_manage_pages
+    )
+    media = media_owner.media.create(alt="alt")
+    variables = {"id": media_global_id(owner_type, media)}
+
+    # when
+    response = staff_api_client.post_graphql(MEDIA_DELETE_MUTATION, variables)
+
+    # then
+    content = get_graphql_content(response)
+    assert content["data"]["mediaDelete"]["errors"] == []
+    mock_media_deleted.assert_called_once_with(media)
+    if owner_type == MediaOwnerTypes.PRODUCT:
+        mock_product_media_deleted.assert_called_once_with(media)
+    else:
+        mock_product_media_deleted.assert_not_called()